registers + function arguments + comments
Comments are with # like in python

you can use register with $ like $rax like normal values, which is
pretty nice for $rdx as the remainder of a division

I have something for function calling conventions. Argumets are
$[argument number], so example $1 for the first argument of the
function. Right now, $0 is argc
and $1 argv for the arguments of the main function